+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Changed
 
+- 注文アイテムとしてのPermitからaccountNumberを削除
+- 決済承認アクションを最適化
+
 ### Deprecated
 
 ### Removed